api-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
httpx==0.27.2
python-dotenv==1.0.1
python-frontmatter==1.1.0
pydantic[email]==2.9.2
